requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0

# Code quality
flake8==6.0.0
black==23.7.0
isort==5.12.0
bandit==1.7.5
safety==2.3.5

# Type checking
mypy==1.4.1
types-requests==2.31.0.2

# Pre-commit hooks
pre-commit==3.3.3

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3 